cleaned up cache service by sperating it into structural and read/write partial classes + moving out all inline classes and enums
