+++ src/CSharp App/Enums/CacheDatabases.cs
@@ -1,0 +1,10 @@
+namespace VolumetricSelection2077.Enums;
+
+public enum CacheDatabases
+{
+    Vanilla,
+    Modded,
+    VanillaBounds,
+    ModdedBounds,
+    All
+}
